# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Rest Sexualidad.",
        "title": "API Rest Sexualidad - Knowdo Studio.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Omar Isalgué Begue",
            "email": "omar26begue@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/images/articles/subtitle/{id}/photo": {
            "get": {
                "description": "Download photo server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Images"
                ],
                "summary": "Download photo server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Image"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/articles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of items.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Returns a list of items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Articles"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create article.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Create article",
                "parameters": [
                    {
                        "description": "Article",
                        "name": "religion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Articles"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/v1/articles/subtitle/{id}/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload an image to the server. The supported extensions are [png, jpeg]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Upload an image to the server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Image Subtitle",
                        "name": "subtitle",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Image"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/v1/doctors/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "It returns Doctor's information.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctors"
                ],
                "summary": "It returns Doctor's information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Articles"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ArticleRequest": {
            "type": "object",
            "required": [
                "category",
                "doctor_article",
                "sub_title",
                "text_article",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "example": "man"
                },
                "doctor_article": {
                    "type": "string",
                    "example": "Omar Isalgue Begue"
                },
                "sub_title": {
                    "type": "string",
                    "example": "Subtitle"
                },
                "text_article": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Title"
                }
            }
        },
        "models.Articles": {
            "type": "object",
            "required": [
                "category",
                "doctor_article",
                "identifier",
                "sub_title",
                "text_article",
                "title"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "date_article": {
                    "type": "string"
                },
                "doctor_article": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "image_article": {
                    "type": "string"
                },
                "image_sub_title": {
                    "type": "string"
                },
                "sub_title": {
                    "type": "string"
                },
                "text_article": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.HTTPResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}